requirements.txt
streamlit
pandas
st-gsheets-connection
pillow
matplotlib
oauth2client